add no result in case query doesn't return content add label in all languages add warning in case category is deleted and used ase filter by the query review the code to build the query
