+++ locales/de.json
@@ -84,6 +84,7 @@
     }
   },
   "query": {
-    "noResult": "Kein Ergebnis für diese Anfrage zurückgegeben"
+    "noResult": "Kein Ergebnis für diese Anfrage zurückgegeben",
+    "catIsMissing": "Eine Kategorie wurde entfernt oder fehlt, bitte überprüfen Sie die Anfrage : '{{queryName}}'"
   }
 }
