+++ locales/en.json
@@ -83,6 +83,7 @@
     }
   },
   "query": {
-    "noResult": "No result returned for this query"
+    "noResult": "No result returned for this query",
+    "catIsMissing": "One category was removed or is missing, please review the query  : '{{queryName}}'"
   }
 }
