+++ locales/es.json
@@ -83,7 +83,8 @@
     }
   },
   "query": {
-    "noResult": "No se devolvió ningún resultado para esta consulta"
+    "noResult": "No se devolvió ningún resultado para esta consulta",
+    "catIsMissing": "Una categoría fue eliminada o está faltante, por favor revise la consulta : '{{queryName}}'"
   }
 
 }
