+++ src/server/views/jcrQuery/JcrQueryInline.jsx
@@ -8,6 +8,7 @@ import {
 } from '@jahia/js-server-core';
 import {HeadingSection} from '../../components';
 import {useTranslation} from 'react-i18next';
+import {buildQuery} from './utils';
 
 export const JcrQueryInline = () => {
     const {t} = useTranslation();
@@ -22,32 +23,21 @@ export const JcrQueryInline = () => {
         'filter',
         'j:subNodesView'
     ]);
-    const asContent = 'content';
-    const descendantPath = luxeQuery.startNode?.getPath() || `/sites/${currentNode.getResolveSite().getSiteKey()}`;
-
-    const filter = luxeQuery.filter?.reduce((condition, categoryNode, index) =>
-        `${condition} ${index === 0 ? '' : 'OR'} ${asContent}.[j:defaultCategory] = '${categoryNode.getIdentifier()}'`
-    , '') || '';
-    const queryFilter = filter.trim().length > 0 ? `AND (${filter})` : '';
-
-    const jcrQuery = `SELECT * FROM [${luxeQuery.type}] AS ${asContent}
-                   WHERE ISDESCENDANTNODE('${descendantPath}')
-                   ${queryFilter}
-                   ORDER BY ${asContent}.[${luxeQuery.criteria}] ${luxeQuery.sortDirection}`;
-
-    server.render.addCacheDependency({flushOnPathMatchingRegexp: `${descendantPath}/.*`}, renderContext);
-
+    const {jcrQuery, warn} = buildQuery({luxeQuery, t, server, currentNode, renderContext});
     const queryContent = getNodesByJCRQuery(currentNode.getSession(), jcrQuery, luxeQuery.maxItems || -1);
 
     return (
         <>
+            {renderContext.isEditMode() && warn &&
+                <p className="text-warning">{warn}</p>}
             {luxeQuery['jcr:title'] &&
                 <HeadingSection title={luxeQuery['jcr:title']}/>}
 
             {queryContent && queryContent.map(node =>
                 <Render key={node.getIdentifier()} node={node} view={luxeQuery['j:subNodesView'] || 'default'}/>
             )}
-            {(!queryContent || queryContent.length === 0) && <em>{t('query.noResult')}</em>}
+            {(!queryContent || queryContent.length === 0) && renderContext.isEditMode() &&
+                <div className="alert alert-dark" role="alert">{t('query.noResult')}</div>}
         </>
     );
 };
